refactor: split VenueCards component into one VenueCard component and one VenueGrid component for reuseability
