+++ src/components/ProfileVenues.jsx
@@ -1,11 +1,13 @@
-import React, { useEffect, useState } from "react";
-import VenueCards from "./VenueCards";
+import { useEffect, useState } from "react";
+import VenueGrid from "./VenueGrid";
+import { useNavigate } from "react-router-dom";
 import { apiRequest } from "../utils/api.mjs";
 import { PROFILE_VENUES } from "../utils/constants.mjs";
 
 export default function ProfileVenues() {
   const [venues, setVenues] = useState([]);
   const [loading, setLoading] = useState(true);
+  const navigate = useNavigate();
 
   const user = JSON.parse(localStorage.getItem("user"));
   const profileName = user.name;
@@ -26,10 +28,28 @@ export default function ProfileVenues() {
 
   return (
     <section className="mt-32">
-      <h1 className="text-3xl font-bold font-heading mb-4 text-center text-copy dark:text-background mb-20">
+      <h1 className="text-3xl font-bold font-heading mb-4 text-center text-copy dark:text-background">
         My venues
       </h1>
-      <VenueCards venues={venues} isOwnerView={true} />
+      <VenueGrid
+        venues={venues}
+        renderFooter={(venue) => (
+          <div className="text-center">
+            <h2 className="text-base font-semibold mb-2">{venue.name}</h2>
+            <div className="flex justify-center gap-8">
+              <button
+                onClick={() => navigate(`/venue/edit/${venue.id}`)}
+                className="bg-copy text-white dark:bg-primary text-copy dark:text-background font-body font-bold px-6 py-1 rounded shadow hover:bg-accent/50 dark:hover:bg-copy hover:text-white transition cursor-pointer"
+              >
+                Edit
+              </button>
+              <button className="bg-copy text-white dark:bg-primary text-copy dark:text-background font-body font-bold px-6 py-1 rounded shadow hover:bg-accent/50 dark:hover:bg-copy hover:text-white transition cursor-pointer">
+                Delete
+              </button>
+            </div>
+          </div>
+        )}
+      />
     </section>
   );
 }
