+++ src/pages/Venues.jsx
@@ -1,5 +1,7 @@
-import React, { useEffect, useState } from "react";
-import VenueCards from "../components/VenueCards";
+import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import VenueGrid from "../components/VenueGrid";
+import { Star } from "lucide-react";
 import Searchbar from "../components/Searchbar";
 import { fetchAllVenues } from "../utils/fetchAllVenues.mjs";
 import { handleError } from "../utils/errorHandler.mjs";
@@ -7,6 +9,7 @@ import { handleError } from "../utils/errorHandler.mjs";
 function Venues() {
   const [venues, setVenues] = useState([]);
   const [loading, setLoading] = useState(true);
+  const navigate = useNavigate();
 
   useEffect(() => {
     fetchAllVenues()
@@ -33,10 +36,32 @@ function Venues() {
     <div>
       <Searchbar />
       <section className="mt-32">
-        <h1 className="text-3xl font-bold font-heading mb-4 text-center text-copy dark:text-background mb-20">
+        <h1 className="text-3xl font-bold font-heading mb-4 text-center text-copy dark:text-background">
           Venues
         </h1>
-        <VenueCards venues={venues} />
+        <VenueGrid
+          venues={venues}
+          renderFooter={(venue) => (
+            <>
+              <div className="flex justify-between">
+                <h2 className="text-base font-semibold">{venue.name}</h2>
+                <div className="flex gap-1 h-8 self-start">
+                  <Star />
+                  <p>{venue.rating}</p>
+                </div>
+              </div>
+              <div className="flex justify-between">
+                <p className="font-thin">${venue.price} one night</p>
+                <button
+                  className="bg-copy text-white dark:bg-primary text-copy dark:text-background font-body font-bold px-2 py-1 rounded shadow hover:bg-accent/50 dark:hover:bg-copy hover:text-white transition cursor-pointer"
+                  onClick={() => navigate(`/venue/${venue.id}`)}
+                >
+                  Read more
+                </button>
+              </div>
+            </>
+          )}
+        />
       </section>
     </div>
   );
